#11 registerページを認証(ログイン)ユーザのみアクセスできるよう設定。editなど認証のみアクセス設定漏れていたところを修正。

+++ routes/web.php
@@ -22,22 +22,23 @@ use Illuminate\Support\Facades\Route;
 Route::get('/', 'PostsController@index');
 // Route::get('/posts/{id}', 'PostsController@show');
 Route::get('/posts/{post}', 'PostsController@show')->where('post', '[0-9]+');
-Route::get('/posts/create', 'PostsController@create');
-Route::post('/posts', 'PostsController@store');
-Route::get('/posts/{post}/edit', 'PostsController@edit');
-Route::patch('/posts/{post}', 'PostsController@update');
-Route::delete('/posts/{post}', 'PostsController@destroy');
+Route::get('/posts/create', 'PostsController@create')->middleware('auth');
+Route::post('/posts', 'PostsController@store')->middleware('auth');
+Route::get('/posts/{post}/edit', 'PostsController@edit')->middleware('auth');
+Route::patch('/posts/{post}', 'PostsController@update')->middleware('auth');
+Route::delete('/posts/{post}', 'PostsController@destroy')->middleware('auth');
 Route::post('/posts/{post}/comments', 'CommentsController@store');
-Route::delete('/posts/{post}/comments/{comment}', 'CommentsController@destroy');
+Route::delete('/posts/{post}/comments/{comment}', 'CommentsController@destroy')->middleware('auth');
+
 
 Route::get('/manage', 'PostsController@manageIndex')->middleware('auth');
-Route::get('/manage/{post}', 'PostsController@manageShow')->where('post', '[0-9]+')->middleware('auth');;
+Route::get('/manage/{post}', 'PostsController@manageShow')->where('post', '[0-9]+')->middleware('auth');
 // where~: パラメータを正規表現でバリデーションしてる(今回は数字かどうかチェック。)
 
+//Good機能
 Route::get('/posts/good/{post}', 'PostsController@increaseGood');
 Route::get('/posts/notgood/{post}', 'PostsController@decreaseGood')->middleware('auth');;
 
-
+//認証機能で追加されたRoute
 Auth::routes();
-
 Route::get('/home', 'HomeController@index')->name('home');
